+++ NEWS.md
@@ -10,6 +10,7 @@
 * Added support for FWE (multiple comparisons p-value adjustment) to `filter_snps()` HWE filtering.
 * Added singleton fitering to `filter_snps()`.
 * Added iPCA `ncp` and `ncp.max` options to `run_random_forest()` and `run_genomic_prediction()`. Previously, selecting the `iPCA` option would work, but run with the default `iPCA` options and thus determine `ncp` internally, which is pretty slow.
+* Added rug plotting to `plot_manhattan()` to allow for easy plotting of gene positions, etc under plots. Both ribbon-style and classic rug style supported.
 
 ## Documentation
 * Fixed some outdated documentation in `calc_ne()`.